feat: debug component
