+++ alca-app/src/components/Debug.js
@@ -1,0 +1,30 @@
+import React from "react";
+import { useSelector } from "react-redux";
+
+import { theme } from "theme/MUITheme";
+import { configuration } from "config/_config";
+
+export default function Debug() {
+    const { reduxState } = useSelector((s) => ({
+        reduxState: s,
+    }));
+
+    // Apply site-wide configs
+    React.useEffect(() => {
+        console.log(`MUI THEME:`, theme);
+        document.title = configuration.site.title;
+    }, []);
+
+    // Setup Debug Print State Key
+    React.useEffect(() => {
+        const printOnD = (e) => {
+            if (e.key === "d") {
+                console.log(reduxState);
+            }
+        };
+        document.addEventListener("keydown", printOnD);
+        return () => document.removeEventListener("keydown", printOnD);
+    });
+
+    return <></>;
+}
